fix: Sync gets post of eventflow (#75)
* fix: sync response body of gets, posts, and cache keys

* fix: active conditions check comparison

* fix: validation of dates provided

* fix: only save cache if the buffer has data

* fix: test script triggering jest

* fix: resolve finally and return conflict
