+++ jest.config.ts
@@ -61,7 +61,7 @@ const config: Config.InitialOptions = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 66,
+      branches: 95,
       functions: 95,
       lines: 95,
       statements: 95,
